+++ CHANGELOG.md
@@ -14,6 +14,6 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/rse-ops/actions-updater/tree/main) (0.0.x)
- - adding tests for updaters (TBA) (0.0.11)
+ - adding tests for updaters, docs, container (0.0.11)
  - adding set-output and set-state updaters (0.0.1)
  - Project skeleton release (0.0.0)